[Windows] refactor stats package to fetch EBS Volume stats for Windows
Presently, the code path in stats package of the agent, which invokes the CSI Client to fetch volume metrics is limited to Linux platform only. The same code path can be used for Windows as well.

This change refactors the existing code to ensure that the EBS volume stats are collected for Windows as well in addition to Linux.
